Finally clean up that mess of counters and garbage

+++ include/graphics/display.hpp
@@ -117,10 +117,10 @@ namespace BlueBear {
           void createWallInstances();
           void updateWallEdgeSegments();
           WallCellBundler& getWallCellBundler( std::shared_ptr< WallCellBundler >& bundlerPtr );
-          std::shared_ptr< XWallInstance > newXWallInstance( int x, int y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper, bool edge = false );
-          std::shared_ptr< YWallInstance > newYWallInstance( int x, int y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper, bool edge = false );
-          std::shared_ptr< DWallInstance > newDWallInstance( int x, int y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper );
-          std::shared_ptr< RWallInstance > newRWallInstance( int x, int y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper );
+          std::shared_ptr< XWallInstance > newXWallInstance( float x, float y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper, bool edge = false );
+          std::shared_ptr< YWallInstance > newYWallInstance( float x, float y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper, bool edge = false );
+          std::shared_ptr< DWallInstance > newDWallInstance( float x, float y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper );
+          std::shared_ptr< RWallInstance > newRWallInstance( float x, float y, float floorLevel, std::string& frontWallpaper, std::string& backWallpaper );
         public:
           void execute();
           void handleEvent( sf::Event& event );
